Add in error trapping.  Fix warnings.  Remove Win8 deprecated calls.  Fix issues with surface thumb rendering.
Also, make updates for changes made to font names and 3rd party file structures.

+++ platform/winrt/mupdf_cpp/PrintPage.cpp
@@ -62,7 +62,7 @@ PrintPages::MakeDocument(IInspectable* doc_options, IPrintDocumentPackageTarget*
 		m_totalpages = m_renderer->GetPrintPageCount();
 
 		for (uint32 page_num = 1; page_num <= m_totalpages; ++page_num)
-			m_renderer->PrintPage(page_num, imageableRect, pageSize, page_desc.DpiX, nullptr);
+			m_renderer->PrintPage(page_num, imageableRect, pageSize, (float) page_desc.DpiX, nullptr);
 	}
 	catch (Platform::Exception^ e)
 	{
